Removed old ItemBuilder
This is a really old class, it was made in Java by someone else (I cannot remember who sorry :( ), It was adapted for KTBukkitGui and written in Kotlin, but I see no need for it since I have never used it. If requested, this change will be reverted but for now use the DSL Item builder instead.